Moves ProxyExchangeHandlerFunction to bean.
This allows easier autoconfiguration.
