+++ spring-cloud-gateway-server-mvc/src/test/java/org/springframework/cloud/gateway/server/mvc/test/LocalServerPortUriResolver.java
@@ -19,6 +19,7 @@ package org.springframework.cloud.gateway.server.mvc.test;
 import java.net.URI;
 
 import org.springframework.cloud.gateway.server.mvc.HandlerFunctions;
+import org.springframework.cloud.gateway.server.mvc.ProxyExchangeHandlerFunction;
 import org.springframework.context.ApplicationContext;
 import org.springframework.web.servlet.function.HandlerFilterFunction;
 import org.springframework.web.servlet.function.HandlerFunction;
@@ -26,7 +27,7 @@ import org.springframework.web.servlet.function.ServerRequest;
 import org.springframework.web.servlet.function.ServerResponse;
 
 public class LocalServerPortUriResolver
-		implements HandlerFunctions.URIResolver, HandlerFilterFunction<ServerResponse, ServerResponse> {
+		implements ProxyExchangeHandlerFunction.URIResolver, HandlerFilterFunction<ServerResponse, ServerResponse> {
 
 	@Override
 	public URI apply(ServerRequest request) {
